corrected: identification of to be deleted package and sig files
Because of an incorrect glob pattern, package files which should not
have been deleted were deleted when removing a package. This happened
for example if the package name contained a "-" (for example, package
files of yay-git were deleted when yay was removed). Now, the
deletion logic uses a regular expression.